minor: PipelineTemplateResouce module opt
