+++ gcloud/template_base/apis/tastypie/resources.py
@@ -1,0 +1,33 @@
+# -*- coding: utf-8 -*-
+"""
+Tencent is pleased to support the open source community by making 蓝鲸智云PaaS平台社区版 (BlueKing PaaS Community
+Edition) available.
+Copyright (C) 2017-2021 THL A29 Limited, a Tencent company. All rights reserved.
+Licensed under the MIT License (the "License"); you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+http://opensource.org/licenses/MIT
+Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+specific language governing permissions and limitations under the License.
+"""
+
+from tastypie.authorization import ReadOnlyAuthorization
+from tastypie.constants import ALL
+
+from pipeline.models import PipelineTemplate
+
+from gcloud.commons.tastypie import GCloudModelResource
+
+
+class PipelineTemplateResource(GCloudModelResource):
+    class Meta(GCloudModelResource.CommonMeta):
+        queryset = PipelineTemplate.objects.filter(is_deleted=False)
+        resource_name = "pipeline_template"
+        authorization = ReadOnlyAuthorization()
+        filtering = {
+            "name": ALL,
+            "creator": ALL,
+            "category": ALL,
+            "subprocess_has_update": ALL,
+            "edit_time": ["gte", "lte"],
+        }
